extract maze representation walls to own namespace
