--- lib/maze_magic/edge.rb
@@ -1,9 +1,0 @@
-module MazeMagic
-  class Edge
-    include Singleton
-
-    def to_console_print
-      ' '
-    end
-  end
-end
