--- lib/maze_magic/horizontal_wall.rb
@@ -1,9 +1,0 @@
-module MazeMagic
-  class HorizontalWall
-    include Singleton
-
-    def to_console_print
-      '_'
-    end
-  end
-end
