+++ lib/maze_magic/maze/edge.rb
@@ -1,0 +1,11 @@
+module MazeMagic
+  module Maze
+    class Edge
+      include Singleton
+
+      def to_console_print
+        ' '
+      end
+    end
+  end
+end
