+++ lib/maze_magic/maze/horizontal_wall.rb
@@ -1,0 +1,11 @@
+module MazeMagic
+  module Maze
+    class HorizontalWall
+      include Singleton
+
+      def to_console_print
+        '_'
+      end
+    end
+  end
+end
