+++ lib/maze_magic/maze/passage.rb
@@ -1,0 +1,11 @@
+module MazeMagic
+  module Maze
+    class Passage
+      include Singleton
+
+      def to_console_print
+        ' '
+      end
+    end
+  end
+end
