+++ lib/maze_magic/maze/vertical_wall.rb
@@ -1,0 +1,11 @@
+module MazeMagic
+  module Maze
+    class VerticalWall
+      include Singleton
+
+      def to_console_print
+        '|'
+      end
+    end
+  end
+end
