--- lib/maze_magic/passage.rb
@@ -1,9 +1,0 @@
-module MazeMagic
-  class Passage
-    include Singleton
-
-    def to_console_print
-      ' '
-    end
-  end
-end
