--- lib/maze_magic/vertical_wall.rb
@@ -1,9 +1,0 @@
-module MazeMagic
-  class VerticalWall
-    include Singleton
-
-    def to_console_print
-      '|'
-    end
-  end
-end
