+++ spec/support/maze_examples.rb
@@ -2,82 +2,82 @@ module MazeExamples
   def cells_grid
     [
       [
-        MazeMagic::Edge.instance,
-        MazeMagic::HorizontalWall.instance,
-        MazeMagic::HorizontalWall.instance,
-        MazeMagic::HorizontalWall.instance,
-        MazeMagic::HorizontalWall.instance,
-        MazeMagic::HorizontalWall.instance,
-        MazeMagic::HorizontalWall.instance,
-        MazeMagic::HorizontalWall.instance,
-        MazeMagic::HorizontalWall.instance,
-        MazeMagic::HorizontalWall.instance,
-        MazeMagic::Edge.instance
+        MazeMagic::Maze::Edge.instance,
+        MazeMagic::Maze::HorizontalWall.instance,
+        MazeMagic::Maze::HorizontalWall.instance,
+        MazeMagic::Maze::HorizontalWall.instance,
+        MazeMagic::Maze::HorizontalWall.instance,
+        MazeMagic::Maze::HorizontalWall.instance,
+        MazeMagic::Maze::HorizontalWall.instance,
+        MazeMagic::Maze::HorizontalWall.instance,
+        MazeMagic::Maze::HorizontalWall.instance,
+        MazeMagic::Maze::HorizontalWall.instance,
+        MazeMagic::Maze::Edge.instance
       ],
       [
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::VerticalWall.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::VerticalWall.instance,
       ],
       [
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::VerticalWall.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::VerticalWall.instance,
       ],
       [
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::VerticalWall.instance
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::VerticalWall.instance
       ],
       [
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::Passage.instance,
-        MazeMagic::VerticalWall.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::Passage.instance,
+        MazeMagic::Maze::VerticalWall.instance,
       ],
       [
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::HorizontalWall.instance,
-        MazeMagic::HorizontalWall.instance,
-        MazeMagic::HorizontalWall.instance,
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::HorizontalWall.instance,
-        MazeMagic::HorizontalWall.instance,
-        MazeMagic::HorizontalWall.instance,
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::HorizontalWall.instance,
-        MazeMagic::VerticalWall.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::HorizontalWall.instance,
+        MazeMagic::Maze::HorizontalWall.instance,
+        MazeMagic::Maze::HorizontalWall.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::HorizontalWall.instance,
+        MazeMagic::Maze::HorizontalWall.instance,
+        MazeMagic::Maze::HorizontalWall.instance,
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::HorizontalWall.instance,
+        MazeMagic::Maze::VerticalWall.instance,
       ]
     ]
   end
@@ -85,14 +85,14 @@ module MazeExamples
   def cells_grid_1by1
     [
       [
-        MazeMagic::Edge.instance,
-        MazeMagic::HorizontalWall.instance,
-        MazeMagic::Edge.instance
+        MazeMagic::Maze::Edge.instance,
+        MazeMagic::Maze::HorizontalWall.instance,
+        MazeMagic::Maze::Edge.instance
       ],
       [
-        MazeMagic::VerticalWall.instance,
-        MazeMagic::HorizontalWall.instance,
-        MazeMagic::VerticalWall.instance
+        MazeMagic::Maze::VerticalWall.instance,
+        MazeMagic::Maze::HorizontalWall.instance,
+        MazeMagic::Maze::VerticalWall.instance
       ]
     ]
   end
